requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
reportlab==4.0.7
gunicorn==22.0.0
django-ratelimit==4.1.0
redis==5.0.1
drf-spectacular==0.27.2

